fix: wgsl-loader and ts-webpack example
The webpack loader was using `schema-utils` without declaring it as a
package dependency, resulting in the version loaded to fall to the
implementation of the package manager. This resulted in an older version
being loaded, which had a different API than the one used in the loader.

This commit declares the dependency with an appropriate version range.

It also cleans up the ts-webpack example, removing the use of the
deprecated Flat component and importing the PropsWithChildren type.