fix: types referencing src directories
The built and published type definitions incorrectly reference the src
directories of the @use-gpu packages. For example, after building the
types, build/ts/layout/src/traits.d.ts will reference core with

```
import("@use-gpu/core/src").ColorLike
```

These paths are incorrect. It appears to be caused by intermixing the TS
composite projects feature with the `paths` option in the tsconfig.

Removing the `paths` appears to fix the issue. I've only tested this so
far with `yarn build` and copying the output to a separate copy of the
ts-webpack example, and verified the types are now correctly loaded.